+++ src/main/java/com/booklink/backend/service/UserService.java
@@ -1,15 +1,14 @@
 package com.booklink.backend.service;
 
-import com.booklink.backend.dto.CreateUserDto;
-import com.booklink.backend.dto.UserDto;
-import com.booklink.backend.dto.UserResponseDto;
+import com.booklink.backend.dto.user.CreateUserDto;
+import com.booklink.backend.dto.user.UserDto;
+import com.booklink.backend.dto.user.UserResponseDto;
 
 import java.util.List;
 
 public interface UserService {
     UserDto registerUser(CreateUserDto userDto);
 
-    UserResponseDto getUser(Long id);
 
     UserResponseDto getUserWithPassword(Long id);
