+++ src/main/java/com/booklink/backend/service/impl/UserServiceImpl.java
@@ -1,9 +1,9 @@
 package com.booklink.backend.service.impl;
 
-import com.booklink.backend.dto.CreateUserDto;
-import com.booklink.backend.dto.UserDto;
-import com.booklink.backend.dto.UserResponseDto;
-import com.booklink.backend.dto.UserWithPasswordDto;
+import com.booklink.backend.dto.user.CreateUserDto;
+import com.booklink.backend.dto.user.UserDto;
+import com.booklink.backend.dto.user.UserResponseDto;
+import com.booklink.backend.dto.user.UserWithPasswordDto;
 import com.booklink.backend.exception.NotFoundException;
 import com.booklink.backend.model.User;
 import com.booklink.backend.repository.UserRepository;
@@ -30,13 +30,6 @@ public class UserServiceImpl implements UserService {
         User userToSave = User.from(userDto, encryptedPassword);
         User savedUser = this.userRepository.save(userToSave);
         return UserDto.from(savedUser);
-    }
-
-    @Override
-    public UserResponseDto getUser(Long id) {
-        Optional<User> userOptional = this.userRepository.findById(id);
-        User user = userOptional.orElseThrow(() -> new NotFoundException("User %d not found".formatted(id)));
-        return UserResponseDto.builder().userDto(UserDto.from(user)).build();
     }
 
     @Override
